backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database & ORM (SQLAlchemy 2.0 style)
sqlalchemy>=2.0.19
asyncpg>=0.28.0
alembic>=1.11.2

# Data Validation & Settings (Pydantic v2)
pydantic>=2.1.1
pydantic-settings>=2.0.2
email-validator>=2.1.0

# Security & Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Web Push Notifications
pywebpush>=2.0.0
py_vapid>=2.0.0
tzdata  # Required on Windows — stdlib zoneinfo has no bundled tz data

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.26.0